refactor(macos): extract Search from AppModel into AppModel+Search.swift
